requirements.txt
asgiref==3.7.2
Authlib==1.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-jwt==1.19.2
idna==3.4
phonenumbers==8.13.18
pycparser==2.21
PyJWT==1.7.1
python-dotenv==0.21.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
